MySQL_API/requirements.txt
altgraph==0.17.2
asgiref==3.5.2
astroid==2.9.3
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.1
bidict==0.22.0
brotlipy==0.7.0
bs4==0.0.1
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cryptography==37.0.4
cycler==0.10.0
Django==4.0.6
dnspython==1.16.0
docopt==0.6.2
Flask-Cors==3.0.9
Flask-Login==0.4.1
Flask-Session==0.3.1
Flask-SocketIO==5.2.0
Flask==1.1.2
fonttools==4.33.3
future==0.18.2
gunicorn==20.0.4
h11==0.13.0
idna==2.10
isort==5.10.1
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
kneed==0.7.0
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
matplotlib==3.5.2
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
mysql-connector==2.2.9
numpy==1.19.5
outcome==1.2.0
packaging==21.3
pandas==1.1.5
pefile==2022.5.30
Pillow==8.1.2
pip==22.2
pipreqs==0.4.11
platformdirs==2.5.1
protobuf==4.21.1
pycparser==2.21
pyinstaller-hooks-contrib==2022.8
pyinstaller==5.2
pylint==2.12.2
pymongo==3.11.3
pyOpenSSL==22.0.0
pyparsing==2.4.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt5-stubs==5.15.6.0
PyQt5==5.15.7
PySocks==1.7.1
python-dateutil==2.8.1
python-engineio==4.3.3
python-socketio==5.7.1
pytz==2021.1
pywin32-ctypes==0.2.0
redis==3.5.3
requests==2.24.0
rq==1.7.0
scipy==1.8.0
selenium==4.3.0
setuptools==47.1.0
six==1.11.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.0.1
sqlparse==0.4.2
toml==0.10.2
trio-websocket==0.9.2
trio==0.21.0
typing-extensions==4.1.1
tzdata==2022.1
urllib3==1.25.11
webdriver-manager==3.3.0
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.13.3
wsproto==1.1.0
yarg==0.1.9
